some changes in category page. Deleted the progress page. and all the functionality of progress to the category.
